Extract PropertiesTypeInspector from ReadablePropertiesTypeInspector and WritablePropertiesTypeInspector + add virtual method CreateProperty to PropertiesTypeInspector and PropertyInfo to ReflectionPropertyDescriptor for #984
